FCM won't work here, read on for detailes
Firebase Cloud Messaging uses the Push API, which means that it needs a
service worker *in all cases*.
https://firebase.google.com/docs/cloud-messaging/js/receive

The firebase code in the sw automagically registers a push message
handler, and
- if the app is in fg:
-	this sw-side fb code will send a MessageEvent (via window.postMessage) to it
- 	the app-side fb code gets it and calls the registered onMessage() handler
- if the app is not in fg:
- 	this sw-side fb code passes the payload to the sw-side onBackgroundMessage() handler

In any case, this is a push notification, and such, it *must* result in showing up
a browser notification. A promise has to be returned from the on*Message() handlers, and
if the Notification API isn't called by the time that promise resolves, Chrome will show
an implicite notification "This site has been updated in the background".

I repeat: in any case. Even for data-only messages.

This whole thing, right from the Push API is a general-purpose one up to this Chrome policy,
so theoretically it could be used for whatever data dispatching we need.

In practice is's usable only for tasks that eventually results in showing up a browser notification.

Which is not our case, so for us it's not the right tool. In fact, anything that uses the Push API
isn't the right tool for us.

So, this code will be cleaned up, but I've spent enough time struggling this mess to keep
it for the slight possibility that we'll need some of it later.